requirements.txt
requests>=2.31.0
PyGitHub>=2.0.0
groq>=0.4.0
python-dotenv>=1.0.0
